+++ src/Frontend/src/pages/settings/Settings.tsx
@@ -1,9 +1,14 @@
+import { Advanced } from "@/components/Advanced";
 import { GithubSettingsPanel } from "./panels/GithubSettingsPanel";
+import { AppSettingsPanel } from "./panels/AppSettingsPanel";
 
 export function Settings() {
 	return (
 		<article class="margin-top">
-			<GithubSettingsPanel />				
+			<Advanced>
+				<GithubSettingsPanel />
+			</Advanced>
+			<AppSettingsPanel />
 		</article>
 	);
 }